Restructure se_shared
