+++ se_apps/include/default_parameters.h
@@ -10,7 +10,7 @@
 #ifndef DEFAULT_PARAMETERS_H_
 #define DEFAULT_PARAMETERS_H_
 
-#include <str_utils.h>
+#include <se/str_utils.h>
 #include <se/utils/math_utils.h>
 #include <vector>
 #include <sstream>
