+++ se_apps/src/benchmark.cpp
@@ -23,7 +23,7 @@
 #include <sstream>
 #include <iomanip>
 #include <getopt.h>
-#include <perfstats.h>
+#include <se/perfstats.h>
 
 PerfStats Stats;
