+++ se_apps/src/reader.cpp
@@ -7,7 +7,7 @@
 
  */
 // #include <DenseSLAMSystem.h>
-#include <str_utils.h>
+#include <se/str_utils.h>
 #include <se/commons.h>
 #include <interface.h>
 #include <se/constant_parameters.h>
