+++ se_denseslam/include/se/DenseSLAMSystem.h
@@ -39,8 +39,8 @@
 #include <se/commons.h>
 #include <iostream>
 #include <memory>
-#include <perfstats.h>
-#include <timings.h>
+#include <se/perfstats.h>
+#include <se/timings.h>
 #include <se/config.h>
 #include <se/octree.hpp>
 #include <se/image/image.hpp>
