+++ se_denseslam/src/DenseSLAMSystem.cpp
@@ -39,8 +39,6 @@
 #include <se/algorithms/meshing.hpp>
 #include <se/geometry/octree_collision.hpp>
 #include <se/vtk-io.h>
-#include "timings.h"
-#include <perfstats.h>
 #include "preprocessing.cpp"
 #include "tracking.cpp"
 #include "rendering.cpp"
