+++ se_denseslam/src/tracking.cpp
@@ -34,7 +34,7 @@
  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. 
  */
 #include <se/utils/math_utils.h>
-#include "timings.h"
+#include <se/timings.h>
 
 #include <se/commons.h>
 #include <se/image/image.hpp>
