+++ se_shared/include/lodepng.h
