+++ se_shared/src/lodepng.cpp
